# timestamp: Switch to a slightly faster datetime_to_timestamp.