@timabbott 2023-02-23
This looks great, merged, thanks @alexmv! I spent a while trying to check what would happen if the parameter was in a non-UTC timezone, before noting we call `verify_UTC` just before this conversion.