@alexmv 2023-02-23
The `verify_UTC` call also takes some time:
```
from zerver.lib.timestamp import verify_UTC
from datetime import datetime, timezone
now = datetime.now(tz=timezone.utc)
%timeit verify_UTC(now)
# 415 ns ± 51.4 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```
...so we could try to make that optional behind a flag and play a little bit fast-and-loose assuming that the fields coming off of this database object are always in UTC, as a further performance optimization.