@timabbott 2023-02-23
Given we call this in 100+ places, I think the extra validation is useful, though maybe we could add an option `skip_utc_check` that can be used for particularly hot code paths if skipping it changes the overall performance for something we care about.